Split admin listings.ts into per-feature modules
listings.ts had grown to ~1050 lines covering create, edit, view,
export, recalculation, lifecycle, uploads and form parsing. Break it
into focused feature modules, leaving listings.ts as a thin route
aggregator (mirroring how attendees.ts assembles its siblings):

- listings-form.ts        form parsing + create/update resources
- listings-uploads.ts     image/attachment upload + delete handlers
- listings-view.ts        detail page + shared attendee helpers
- listings-export.ts      attendee CSV export
- listings-edit.ts        create / duplicate / edit
- listings-recalculate.ts aggregate recalculation
- listings-lifecycle.ts   deactivate / reactivate / delete / log

To keep the 0% duplication threshold, the question-answer loading and
date-filtered attendee handling shared by the detail page and the CSV
export are extracted into loadListingQuestionData and
filteredAttendeesHandler, and the duplicate/edit GET handlers share a
listingAndGroupsPage helper. No behaviour change.
